Merged PR 1159991: [Infrastructure] Add support for configuration, and update the CI pipeline
### Description:

Configuration works differently to what we have in Share. The configuration is specified in `templates/appsettings/appsettings.json`. The contents of `appsettings.json` gets injected as a JSON string into `scopeBindings.json` during the build. The `$config(...)` sections are filled in during the deployment. The JSON string is available to our app through an environment variable. I added a sample configuration to show how this works. The DI code for our configuration needs some work, but I want to commit it as is to unblock others.

Other miscellaneous changes:

- Update the CI pipeline to build a docker image
- Remove a step from the official pipeline that was used for debugging

### PR Checklist:
- [x] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [x] Infrastructure/Configuration