Merged PR 1169788: Create PowerBI Credential
### Description:
A PowerBI credential holds the login/password for connecting to a SQL-based data source. The credential will need to be created in Synapse.

Each Purview account will have a unique login to isolate the database schema. To do this, we hash the accountId.

All Purview accounts will share the same password stored in AKV. This was reviewed with Platform security (V.B.) and signed off. The health service will be responsible for generating the random password. To safely and securely maintain the password, we will push it into AKV.

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [x] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [x] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [x] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Related work items: #2725558