Merged PR 1173266: Assign workspace to a created capacity during provisioning
### Description:
Assign workspace to a created capacity during provisioning. Wire up the account creation API path **only** for local development access, otherwise return 200 with no execution

The expectation is that some PBI capacities exist (infra), and during provisioning, we must assign a workspace to that capacity before we can create a report. The lookup of the capacities is currently hard coded, but the expectation is to do a lookup of the capacities dynamically. An exposure control is used to manage this so we can provision, scale and adjust capacities easily for prod resources.

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [x] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [x] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
![image.png](https://msdata.visualstudio.com/4031b34e-6354-4257-94de-a85346a777ae/_apis/git/repositories/7bb12a8c-53e2-492d-ae1c-d1ba95b637f1/pullRequests/1173266/attachments/image.png)

Related work items: #2738917