Merged PR 1175101: Create a shared dataset
### Description:
A shared dataset is one which is referenced by another report using a live connection. We will have a shared dataset for CDMC and create the visual report with a reference to it. To accomplish this, first we import a dataset for the shared CDMC then remove the linked report associated as it is not needed. Then we create the visual CDMC report and finally bind that report to the shared dataset.

- abstract the report operations from health report component
- add service operations to rebind and delete a report
- move the server properties to a dataset request property as it is not always used when importing a dataset

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [x] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [x] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)

Related work items: #2746017