Merged PR 1217575: Trends Controller + Set up for Queries
Credit to: Priyanka Chadalavada
### Description:
This PR is intended to get Priyanka's changes into a working state and merged in an effort to release incremental PRs for finishing up the Trends APIs.

It implements the Trends controller and provides the set up for implementing the HealthTrendRepository and queries.

Note: I stubbed out the `HealthTrendRepository.GetSingle()` method, which triggers ResourceNotFound exception in the `HealthTrendComponent.Get()` method. At the moment, nothing is calling these two APIs, so it should be fine.

My next PR will include the queries for trends, and is almost done. Will release some time tomorrow

### PR Checklist:
- [X] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [X] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [X] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Related work items: #2803861, #2816548