Merged PR 1222575: Provisioning service - add out of the box control entities
### Description:
We have 11 out of the box controls, some of which are "Data governance score", "Metadata completeness", "Use". When an account is provisioned, we need to persist the out of box controls into the artifact store. That way, when the UX calls GET /controls, we can grab the information necessary from the artifact store and return it.

When GET /controls is called, all properties in the response are populated from the artifact store except two; "CurrentScore", "LastRefreshedAt" which are populated from synapse.

Successful test logs:
https://portal.microsoftgeneva.com/s/9D0BEE87

2 existing records and the process inserted the remaining 9 records.
![image.png](https://msdata.visualstudio.com/4031b34e-6354-4257-94de-a85346a777ae/_apis/git/repositories/7bb12a8c-53e2-492d-ae1c-d1ba95b637f1/pullRequests/1222575/attachments/image.png)

### PR Checklist:
- [X] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [X] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [X] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Related work items: #2847020