Merged PR 1226591: Refresh PBI Report in background
### Description:
Create a refresh callback that will initiate a refresh for all system datasets. Monitor the progress until all of the dataset refresh are in a terminal status (success or fail).

Initialize the dependencies of the worker job on startup and copy the needed configurations into the appsettings.

Has not been wired into the callback of the spark job to initiate so it is a no-op right now

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [x] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [x] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

Related work items: #2854339