Merged PR 1233678: remove account id from data access event
Common event hub model contains the AccountId property which has been duplicated in DataAccess schema. This is now removed from DataAcces

----
#### AI description (iteration 1)
This pull request includes a modification in the **DataAccessEventHubEntityModels.cs** file. Specifically, a line of code has been removed from the `GetSchemaDefinition` method in the `PolicySetEventHubEntityModel` class. The removed line was responsible for creating a new `StructField` named "AccountId". This suggests that the "AccountId" field may no longer be necessary or relevant in this context.