Merged PR 1243212: Populate StackTrace and ex. message columns when logging error, critical and warning
### Description:
_<Describe proposed changes here...>_

### PR Checklist:
- [X] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

add changes

----
#### AI description (iteration 1)
The pull request introduces changes in the **DataEstateHealthLogger.cs** file, specifically in the logging methods:
- `LogWarning`
- `LogError`
- `LogCritical`

In each of these methods, the call to `GetAdditionalColumns` now includes an additional parameter for exceptions (`exception`). This change enhances the logging functionality by including exception details in the logs. Additionally, the method signature of `GetAdditionalColumns` has been updated to accept this new parameter. The exception details are then added as a key-value pair to the list of additional columns that are logged. This ensures that any exceptions occurring during these log levels will be captured and logged for further analysis.

Related work items: #2881519