Merged PR 1247981: set aks config prior to reference of service config
### Description:
set aks config prior to reference of service config

null reference occurred during logger setup when having aks config within kestrel setup. Configurations must be configured before hand and only required for production (non-local) use.

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [x] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

set aks config prior to reference of service config

----
#### AI description (iteration 1)
#### PR Classification
This pull request is a code refactoring change aimed at improving the configuration setup for the application in non-development environments.

#### PR Summary
The pull request modifies the configuration setup process in the `Program.cs` file of the `ApiService` project. The changes ensure that the AKS configuration is set only when the environment is not in development mode.
- In `Program.cs`, the `SetAksConfiguration(builder)` method call has been moved from the `ConfigureKestrelServerForProduction` method to the main class constructor, and it is now wrapped in a conditional statement to check if the environment is not in development mode.
- The `SetAksConfiguration(builder)` method call has been removed from the `ConfigureKestrelServerForProduction` method.