Merged PR 1279907: Fix cert
#### AI description (iteration 1)
#### PR Classification
This pull request is a configuration change aimed at fixing an issue with the certificate.

#### PR Summary
The pull request modifies the `dataQualitySubjectName` configuration in two JSON files, changing the prefix from `pdg.environment.shortName` to `regionShortName`. This change is likely to affect the way certificates are identified or accessed in the system.
- In `config.prod.Microsoft.Purview.PDG.DataEstateHealth.json`, the `dataQualitySubjectName` configuration was updated.
- Similarly, in `config.test.Microsoft.Purview.PDG.DataEstateHealth.json`, the `dataQualitySubjectName` configuration was also updated.