Merged PR 1290128: Fix template
#### AI description (iteration 1)
#### PR Classification
This pull request is a minor bug fix, addressing inconsistencies in the naming and rule checking within the `CDMC.json` template.

#### PR Summary
The changes in this pull request correct the capitalization of "Data cataloging" and update a rule check point in the `CDMC.json` file. These modifications ensure consistency and accuracy in the template.
- In `/src/BusinessLogic/Templates/ControlAssessment/CDMC.json`, the control and assessment name "Data Cataloging" has been changed to "Data cataloging".
- In the same file, the rule check point "DataProductAllRelatedAssetsHaveOwner" has been updated to "DataProductAllRelatedAssetsHaveDQScore".