Merged PR 1306329: Update template - action name for "Missing published glossary terms on data products"
#### AI description (iteration 1)
#### PR Classification
This pull request is a minor update to the action names in the CDMC.json template file.

#### PR Summary
The pull request updates the action names in the CDMC.json file to better reflect the intent of the actions. The changes are aimed at improving clarity and understanding of the action's purpose.
- Updated the action name from "Missing related terms on critical data entities" to "Missing published glossary terms on critical data entities" in the CDMC.json file.
- The action name "Missing related terms on data products" was changed to "Missing published glossary terms on data products" in three different sections of the CDMC.json file.