Merged PR 1311731: Remove the linked service dependency in domain model and move to cosmos DB read only key.
### Description:
Remove the linked service dependency in domain model and move to cosmos DB read only key.

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [X] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [X] Infrastructure/Configuration

### Additional Details (optional)
This is interim fix, to avoid accessing the environment eventually move to user managed identity when spark supports reading from cosmosdb analytical store.

Remove the linked service dependency and move to cosmos db read only key..

----
#### AI description (iteration 1)
#### PR Classification
This pull request involves a design change, specifically the removal of a linked service dependency in the domain model and a shift to using a Cosmos DB read-only key.

#### PR Summary
The pull request modifies the `CatalogSparkJobComponent` class in the `CatalogSparkJobComponent.cs` file to remove the dependency on a linked service and instead use a Cosmos DB read-only key for data access.
- In `CatalogSparkJobComponent.cs`, the constructor now includes an `IKeyVaultAccessorService` parameter, which is used to retrieve the Cosmos DB name and read-only key.
- The `SubmitJob` method in `CatalogSparkJobComponent.cs` now retrieves the Cosmos DB name and key from the Key Vault.
- The `GetSparkJobRequest` and `GetSinkConfiguration` methods in `CatalogSparkJobComponent.cs` have been updated to include the Cosmos DB name and key in their configurations.
- The `CatalogSparkJobComponent.cs` file also has some minor changes in its imports, removing unused ones and adding the `Azure.Security.KeyVault.Secrets` namespace.