Merged PR 1334122: Increase timeout and enhance log
#### AI description (iteration 1)
#### PR Classification
This pull request primarily focuses on enhancing error handling and logging, as well as increasing timeout durations in the Data Quality Services.

#### PR Summary
The changes in this pull request aim to improve the robustness and traceability of the Data Quality Services by introducing specific exceptions and enhancing logging. Additionally, the timeout durations have been increased to accommodate longer running operations.
- In `DataQualityExecutionService.cs`, error handling has been added to `ParseDQResult` and `PurgeObserver` methods, throwing `ParseMDQResultException` and `PurgeObseverException` respectively in case of any exceptions.
- Two new exception classes `ParseMDQResultException.cs` and `PurgeObserverException.cs` have been introduced to provide more specific error information.
- In `DataQualityOutputRepository.cs`, the `DefaultTimeout` and `DefaultConnectTimeout` constants have been increased from 10 minutes to 20 minutes.
- In `DataQualityServiceClientFactory.cs`, the HttpClient timeout has been set to 5 minutes.