Merged PR 1354674: Skip provision DH if EC is not enabled
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature that allows the system to skip the provisioning of Data Governance Health (DH) if it is not enabled.

#### PR Summary
The changes in this pull request introduce a conditional check to determine if Data Governance Health provisioning is enabled before proceeding with the provisioning process. If it is not enabled, the system logs the information and returns an OK response. This feature is implemented to prevent unnecessary provisioning when the feature is not enabled.
- In `PlatformAccountNotificationsController.cs`, a conditional check is added to verify if Data Governance Health provisioning is enabled before proceeding with the provisioning tasks.
- In `AccountExposureControlConfigProvider.cs`, a new method `IsDataGovHealthProvisioningEnabled` is added to check if the feature is enabled.
- In `IAccountExposureControlConfigProvider.cs`, the interface is updated to include the new method `IsDataGovHealthProvisioningEnabled`.