Merged PR 1379840: Check if processing storage account mapping exists
#### AI description  (iteration 1)
#### PR Classification
This pull request introduces a new feature to handle the scenario when a processing storage account mapping does not exist.

#### PR Summary
The changes in this pull request add exception handling for the case when a processing storage account mapping is not found. This is done by introducing a new exception class and modifying the existing services to throw and catch this new exception.
- In `DHScheduleService.cs`, the catch block is updated to handle the new `ProcessingStorageAccountMappingNotExistsException` and update the job status to failed.
- In `DataQualityExecutionService.cs`, a check is added to throw the new `ProcessingStorageAccountMappingNotExistsException` if the account storage model is null.
- A new exception class `ProcessingStorageAccountMappingNotExistsException.cs` is added to represent the scenario when a processing storage account mapping does not exist.